+++ charmike-frontend/src/context/AuthContext.jsx
@@ -1,5 +1,5 @@
 import { createContext, useContext, useEffect, useState } from "react";
-import * as authService from "../services/auth";
+import authService from "@/services/auth";
 
 const AuthContext = createContext();